feat: enable strict compilation with maximum warnings as errors
- Add MORPH_ENABLE_STRICT_COMPILATION option (default ON)
- Enable -Werror and ~60 additional warning flags across GCC, Clang, MSVC
- Move shared diagnostics (Wshadow, -Wconversion, etc.) from Clang-only
  to the common GNU/Clang baseline
- Fix all warnings surfaced by the new flags:
  - Add missing default: labels in switch statements
  - Remove empty @par Example paragraphs from doxygen docs
  - Suppress -Wfloat-equal in tests with intentional float comparisons
  - Mark bridge-registered helper methods as [[maybe_unused]]
  - Suppress GCC -Wpedantic on __int128 extension usage
  - Remove redundant static_cast on size_t in session_auth.hpp
  - Add missing .roles and .idempotencyKey fields in designated
    initializers

+++ include/morph/session_auth.hpp
@@ -58,7 +58,7 @@ inline std::string sha256(std::string_view data) {
                                     0x510e527f, 0x9b05688c, 0x1f83d9ab, 0x5be0cd19};
 
     std::vector<uint8_t> msg(data.begin(), data.end());
-    const uint64_t bitLen = static_cast<uint64_t>(msg.size()) * 8;
+    const uint64_t bitLen = msg.size() * 8;
     msg.push_back(0x80);
     while (msg.size() % 64 != 56) {
         msg.push_back(0x00);
